Update README.md
This video explains a power supply cut-off system for an ancillary device such as an Arduino module, pressing a BP arm the power supply, pressing another BP disarms this power supply, the BPs can be replaced with pulse-type signals from an MCU, for example, so an MCU powered by this device can decide for itself to cut itself off from its power supply by sending a rising edge signal to instead of the disarming PB this power supply can be awakened to external signals , I have sometimes called this system power saving, because it allows to limit the discharge of the battery or battery, when it is not necessary. To give an example, the awakening of the power supply could for example a tilt contact, so by shaking the device, this would allow it to be awakened.